// src/app/interfaces/auth.ts
// Generated by https://quicktype.io

export interface Auth {
  user: User;
}

export interface User {
  id: string;
  first_name: string;
  last_name: string;
  email: string;
  role: string;
  access_token: string;
}

export interface userRegister {
  first_name: string;
  last_name: string;
  email: string;
  password: string;
  type: number;
}